+++ src/inmemorystorage.ts
@@ -1,4 +1,4 @@
-import EventHandling from './eventhandling-new';
+import EventHandling from './eventhandling';
 import CachingLayer from './cachinglayer';
 import { applyMixins } from './util';
